refactor(engine): 拆分 Claude 事件转换模块并完成大文件治理
本次提交以无行为变更为前提，对 Claude 引擎事件转换逻辑进行模块化拆分，目标是将超阈值文件回落到治理线以内。

主要改动：

- 将 convert_event 与 convert_stream_event 从 claude.rs 迁移到 src-tauri/src/engine/claude/event_conversion.rs

- 在 claude.rs 增加 #[path = "claude/event_conversion.rs"] mod event_conversion，确保 cc_gui_daemon 通过 #[path] 引入时仍能正确解析子模块

- 精简 claude.rs 的 stream_helpers 导入，extract_tool_result_text 仅在测试场景下通过 #[cfg(test)] 导入，避免非测试构建中的 unused import 噪音

兼容性与行为说明：

- 保持 ClaudeSession 对外契约不变，不修改命令参数、事件类型与返回语义

- ToolStarted/ToolCompleted/ToolOutputDelta、ReasoningDelta、TextDelta 等事件映射流程保持原实现语义

- AskUserQuestion 相关的 request/恢复链路保持不变

验证结果：

- cargo check --manifest-path src-tauri/Cargo.toml 通过

- cargo test --manifest-path src-tauri/Cargo.toml engine::claude -- --nocapture 通过（engine::claude 相关 52 项测试通过）

- node scripts/check-large-files.mjs 输出 threshold>3000, found=0
